refactor: Main -> Controller
